Backend/requirements.txt
asgiref==3.11.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
#python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.5
